+++ integration/spring-data/2.1/src/main/java/com/blazebit/persistence/spring/data/impl/query/PartTreeBlazePersistenceQuery.java
@@ -30,6 +30,7 @@ import com.blazebit.persistence.view.EntityViewManager;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.provider.PersistenceProvider;
 import org.springframework.data.jpa.repository.query.AbstractJpaQuery;
+import org.springframework.data.jpa.repository.query.EscapeCharacter;
 import org.springframework.data.jpa.repository.query.JpaQueryExecution;
 import org.springframework.data.repository.query.ParameterAccessor;
 import org.springframework.data.repository.query.Parameters;
@@ -38,6 +39,7 @@ import org.springframework.data.repository.query.parser.PartTree;
 
 import javax.persistence.EntityManager;
 import javax.persistence.Query;
+import javax.persistence.criteria.CriteriaBuilder;
 import java.util.Collections;
 import java.util.List;
 
@@ -49,8 +51,18 @@ import java.util.List;
  */
 public class PartTreeBlazePersistenceQuery extends AbstractPartTreeBlazePersistenceQuery {
 
-    public PartTreeBlazePersistenceQuery(EntityViewAwareJpaQueryMethod method, EntityManager em, PersistenceProvider persistenceProvider, CriteriaBuilderFactory cbf, EntityViewManager evm) {
-        super(method, em, persistenceProvider, cbf, evm);
+    public PartTreeBlazePersistenceQuery(EntityViewAwareJpaQueryMethod method, EntityManager em, PersistenceProvider persistenceProvider, EscapeCharacter escape, CriteriaBuilderFactory cbf, EntityViewManager evm) {
+        super(method, em, persistenceProvider, escape, cbf, evm);
+    }
+
+    @Override
+    protected ParameterMetadataProvider createParameterMetadataProvider(CriteriaBuilder builder, ParametersParameterAccessor accessor, PersistenceProvider provider, Object escape) {
+        return new ParameterMetadataProviderImpl(builder, accessor, provider, (EscapeCharacter) escape);
+    }
+
+    @Override
+    protected ParameterMetadataProvider createParameterMetadataProvider(CriteriaBuilder builder, JpaParameters parameters, PersistenceProvider provider, Object escape) {
+        return new ParameterMetadataProviderImpl(builder, parameters, provider, (EscapeCharacter) escape);
     }
 
     @Override